extract inner class from LinkedDecisionDiagram
