+++ output/MCP_exact.dot
@@ -1,0 +1,31 @@
+digraph exact{
+0 [label="[0, 0, 0, 0]
+fub: 6 - value: 0", style=rounded, shape=rectangle, tooltip="Exact node"];
+1 [label="[0, 1, 2, -2]
+fub: 8 - value: 0", style=rounded, shape=rectangle, tooltip="Exact node"];
+3 [label="[0, 0, -1, -1]
+fub: 2 - value: 1", style=rounded, shape=rectangle, tooltip="Exact node"];
+2 [label="[0, 0, 5, -3]
+fub: 8 - value: 0", style=rounded, shape=rectangle, tooltip="Exact node"];
+4 [label="[0, 0, 0, -2]
+fub: 2 - value: 5", style=rounded, shape=rectangle, tooltip="Exact node"];
+6 [label="[0, 0, 0, -4]
+fub: 4 - value: 0", style=rounded, shape=rectangle, tooltip="Exact node"];
+5 [label="[0, 0, 0, 0]
+fub: 0 - value: 1", style=rounded, shape=rectangle, tooltip="Exact node"];
+7 [label="[0, 0, 0, 0]
+fub: 0 - value: 7", style=rounded, shape=rectangle, tooltip="Exact node", style="filled, rounded", shape=rectangle, color=black, fontcolor=white, tooltip="Terminal node"];
+6 -> 7 [label=4, tooltip="Decision: assign 0 to var 3"];
+2 -> 6 [label=0, tooltip="Decision: assign 0 to var 2"];
+3 -> 5 [label=0, tooltip="Decision: assign 1 to var 2"];
+0 -> 1 [label=0, tooltip="Decision: assign 0 to var 0", color="#6fb052", fontcolor="#6fb052"];
+1 -> 2 [label=0, tooltip="Decision: assign 0 to var 1", color="#6fb052", fontcolor="#6fb052"];
+4 -> 7 [label=0, tooltip="Decision: assign 1 to var 3"];
+1 -> 3 [label=1, tooltip="Decision: assign 1 to var 1"];
+2 -> 4 [label=5, tooltip="Decision: assign 1 to var 2", color="#6fb052", fontcolor="#6fb052"];
+6 -> 7 [label=0, tooltip="Decision: assign 1 to var 3"];
+5 -> 7 [label=0, tooltip="Decision: assign 0 to var 3"];
+5 -> 7 [label=0, tooltip="Decision: assign 1 to var 3"];
+3 -> 4 [label=1, tooltip="Decision: assign 0 to var 2"];
+4 -> 7 [label=2, tooltip="Decision: assign 0 to var 3", color="#6fb052", fontcolor="#6fb052"];
+}
