+++ src/main/java/org/ddolib/ddo/core/mdd/NodeType.java
@@ -1,0 +1,8 @@
+package org.ddolib.ddo.core.mdd;
+
+/**
+ * Flag to identify the type of node: exact node, relaxed node, marked node, etc ...
+ */
+enum NodeType {
+    EXACT, RELAXED
+}
